Add indexIdentifier parameter to put function in ValueStore
This update adds an 'indexIdentifier' parameter to the 'put' function in the ValueStore class, allowing the logging of which application the updates come from. This will provide more detailed tracking data in our value stores and enhance data accuracy. The App logic is also updated to pass this new parameter when calling the put function.